move more stuff out of main
